code push with error

+++ src/app/modules/user/user.service.ts
@@ -2,6 +2,9 @@ import { User } from "../user.model";
 import { TUser } from "./user.interface";
 
 const createUserIntoDB=async(userData:TUser)=>{
+     if (await User.isUserExists(userData.userId)) {
+       throw new Error("User already exists");
+     }
 const result=await User.create(userData)
 return result;
 }
@@ -11,19 +14,26 @@ const getUsersDataFromDB =async()=>{
     return result;
 }
 
-const getSingleUserFromDB=async(id:string)=>{
-    console.log(id);
+const getSingleUserFromDB=async(id:number)=>{
+   if (!(await User.isUserExists(id))) {
+     throw new Error("User not found");
+   }
     const result=await User.findOne({userId:id});
-    console.log(result)
     return result;
 }
 
-const updateUserToDB = async (id: string, updateData: TUser) => {
+const updateUserToDB = async (id: number, updateData: TUser) => {
+    if (!(await User.isUserExists(id))) {
+      throw new Error("User not found");
+    }
   const result = await User.replaceOne({ userId: id }, updateData);
   return result;
 };
 
-const deleteUserFromDB= async(id: string)=>{
+const deleteUserFromDB= async(id: number)=>{
+    if (!(await User.isUserExists(id))) {
+      throw new Error("User not found");
+    }
     const result = await User.deleteOne({userId:id});
     return result;
 }
